Corrected indentation of code block under Load Data into a Dataframe step 10
Corrected indentation of the code lines under "Load Data into a Dataframe" step 10, where the code lines were one space too short due to the addition of the additional character in "10.". 

This caused the code block in step 10 to not be interpreted as a code block and instead show as plain text, and the step numbering for step 11 to incorrectly show as step 1.