Update of files PBIX in the "Starter" and "Solution" folders
Allfiles/Lab/15/Starter
Allfiles/Lab/15/Solution

The PBIX files in the "Starter" and "Solution" folders are a different version from the files inside the "15-scalable-semantic-models.zip" zip file.

The PBIX version in the "Starter" folder is outdated (the "Total Sales" column does not exist in the "Sales" table, and the "Year" column does not exist in the "Date" table).